third_party/WebKit/LayoutTests/inspector/domdebugger/domdebugger-getEventListeners.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>
function test()
{
    var windowObject;
    var divWithHandlers;
    var divWithoutHandlers;
    function dumpListeners(next, listeners)
    {
        listeners.sort((a, b) => a.type().localeCompare(b.type()));
        for (var i = 0; i < listeners.length; ++i) {
            delete listeners[i]._location.scriptId;
            var sourceURL = listeners[i]._sourceURL;
            sourceURL = sourceURL.substr(sourceURL.lastIndexOf('/') + 1);
            listeners[i]._sourceURL = sourceURL;

            InspectorTest.addResult("type: " + listeners[i].type());
            InspectorTest.addResult("useCapture: " + listeners[i].useCapture());
            InspectorTest.addResult("location: " + listeners[i].location().columnNumber + ", " + listeners[i].location().lineNumber);
            InspectorTest.addResult("handler: " + listeners[i].handler().description);
            InspectorTest.addResult("sourceURL: " + listeners[i].sourceURL());
            InspectorTest.addResult("");
        }
        next();
    }

    InspectorTest.runTestSuite([
        function testSetUp(next)
        {
            InspectorTest.RuntimeAgent.evaluate("window", "get-event-listeners-test", step1);

            function step1(error, result, exceptionDetails)
            {
                windowObject = InspectorTest.runtimeModel.createRemoteObject(result);
                InspectorTest.RuntimeAgent.evaluate("document.getElementById(\"with-handlers\")", "get-event-listeners-test", step2);
            }

            function step2(error, result, exceptionDetails)
            {
                divWithHandlers = InspectorTest.runtimeModel.createRemoteObject(result);
                InspectorTest.RuntimeAgent.evaluate("document.getElementById(\"without-handlers\")", "get-event-listeners-test", step3);
            }

            function step3(error, result, exceptionDetails)
            {
                divWithoutHandlers = InspectorTest.runtimeModel.createRemoteObject(result);
                next();
            }
        },

        function testWindowEventListeners(next)
        {
            windowObject.eventListeners().then(dumpListeners.bind(this, next));
        },

        function testDivEventListeners(next)
        {
            divWithHandlers.eventListeners().then(dumpListeners.bind(this, next));
        },

        function testDivWithoutEventListeners(next)
        {
            divWithoutHandlers.eventListeners().then(dumpListeners.bind(this, next));
        }
    ]);
}

</script>
</head>
<body>
<p>
Tests RemoteObject.eventListeners.
</p>
<div id="with-handlers" onclick="return 42;"></div>
<div id="without-handlers"></div>
<script>
    function foo() {}
    function boo() {}
    window.addEventListener("scroll", foo, true);
    document.getElementById("with-handlers").addEventListener("click", boo, true);
    document.getElementById("with-handlers").addEventListener("mouseout", foo, false);
    runTest();
</script>
</body>
</html>
